+++ src/main/java/org/truffleruby/core/regexp/RegexpNodes.java
@@ -13,11 +13,13 @@ import java.util.Arrays;
 import java.util.Iterator;
 
 import com.oracle.truffle.api.dsl.Bind;
+import com.oracle.truffle.api.dsl.NeverDefault;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.profiles.InlinedBranchProfile;
 import com.oracle.truffle.api.strings.TruffleString;
 import org.joni.NameEntry;
 import org.truffleruby.annotations.CoreMethod;
+import org.truffleruby.annotations.Split;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.annotations.CoreModule;
 import org.truffleruby.annotations.Primitive;
@@ -25,21 +27,23 @@ import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
 import org.truffleruby.core.array.RubyArray;
 import org.truffleruby.core.cast.ToStrNode;
 import org.truffleruby.core.encoding.Encodings;
+import org.truffleruby.core.encoding.RubyEncoding;
 import org.truffleruby.core.encoding.TStringUtils;
 import org.truffleruby.core.klass.RubyClass;
 import org.truffleruby.core.string.ATStringWithEncoding;
+import org.truffleruby.core.string.StringHelperNodes;
 import org.truffleruby.core.string.TStringWithEncoding;
 import org.truffleruby.core.string.RubyString;
 import org.truffleruby.core.symbol.RubySymbol;
 import org.truffleruby.annotations.Visibility;
+import org.truffleruby.language.PerformanceWarningNode;
 import org.truffleruby.language.control.DeferredRaiseException;
 import org.truffleruby.language.control.RaiseException;
 import org.truffleruby.language.library.RubyStringLibrary;
 
-import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Cached;
-import com.oracle.truffle.api.dsl.Fallback;
+import com.oracle.truffle.api.dsl.Cached.Shared;
 import com.oracle.truffle.api.dsl.Specialization;
 
 @CoreModule(value = "Regexp", isClass = true)
@@ -54,41 +58,67 @@ public abstract class RegexpNodes {
         }
     }
 
-    @CoreMethod(names = { "quote", "escape" }, onSingleton = true, required = 1)
+    @CoreMethod(names = { "quote", "escape" }, onSingleton = true, required = 1, split = Split.ALWAYS)
     public abstract static class QuoteNode extends CoreMethodArrayArgumentsNode {
-
-        @Child private QuoteNode quoteNode;
 
         public abstract RubyString execute(Object raw);
 
+        @NeverDefault
         public static QuoteNode create() {
             return RegexpNodesFactory.QuoteNodeFactory.create(null);
         }
 
-        @Specialization(guards = "libRaw.isRubyString(raw)", limit = "1")
+        @SuppressWarnings("truffle-static-method")
+        @Specialization(
+                guards = {
+                        "libString.isRubyString(raw)",
+                        "equalNode.execute(node, libString, raw, cachedString, cachedEnc)" },
+                limit = "getDefaultCacheLimit()")
+        RubyString quoteStringCached(Object raw,
+                @Cached @Shared RubyStringLibrary libString,
+                @Cached("asTruffleStringUncached(raw)") TruffleString cachedString,
+                @Cached("libString.getEncoding(raw)") RubyEncoding cachedEnc,
+                @Cached StringHelperNodes.EqualSameEncodingNode equalNode,
+                @Bind("this") Node node,
+                @Cached("quote(libString, raw)") TStringWithEncoding quotedString) {
+            return createString(quotedString);
+        }
+
+        @Specialization(replaces = "quoteStringCached", guards = "libString.isRubyString(raw)")
         RubyString quoteString(Object raw,
-                @Cached RubyStringLibrary libRaw) {
-            return createString(ClassicRegexp.quote19(new ATStringWithEncoding(libRaw, raw)));
+                @Cached @Shared RubyStringLibrary libString) {
+            return createString(quote(libString, raw));
         }
 
-        @Specialization
+        @Specialization(guards = "raw == cachedSymbol", limit = "getDefaultCacheLimit()")
+        RubyString quoteSymbolCached(RubySymbol raw,
+                @Cached("raw") RubySymbol cachedSymbol,
+                @Cached("quote(cachedSymbol)") TStringWithEncoding quotedString) {
+            return createString(quotedString);
+        }
+
+        @Specialization(replaces = "quoteSymbolCached")
         RubyString quoteSymbol(RubySymbol raw) {
-            return doQuoteString(createString(raw.tstring, raw.encoding));
+            return createString(quote(raw));
         }
 
-        @Fallback
-        RubyString quote(Object raw,
-                @Cached ToStrNode toStrNode) {
-            return doQuoteString(toStrNode.execute(this, raw));
+        @Specialization(guards = { "!libString.isRubyString(raw)", "!isRubySymbol(raw)" })
+        static RubyString quoteGeneric(Object raw,
+                @Cached @Shared RubyStringLibrary libString,
+                @Cached ToStrNode toStrNode,
+                @Cached QuoteNode recursive,
+                @Bind("this") Node node) {
+            return recursive.execute(toStrNode.execute(node, raw));
         }
 
-        private RubyString doQuoteString(Object raw) {
-            if (quoteNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                quoteNode = insert(QuoteNode.create());
-            }
-            return quoteNode.execute(raw);
+        TStringWithEncoding quote(RubyStringLibrary strings, Object string) {
+            return ClassicRegexp.quote19(new ATStringWithEncoding(strings, string));
         }
+
+        TStringWithEncoding quote(RubySymbol symbol) {
+            return ClassicRegexp.quote19(new ATStringWithEncoding(symbol.tstring, symbol.encoding));
+        }
+
     }
 
     @CoreMethod(names = "source")
@@ -180,12 +210,39 @@ public abstract class RegexpNodes {
     @Primitive(name = "regexp_compile", lowerFixnum = 1)
     public abstract static class RegexpCompileNode extends PrimitiveArrayArgumentsNode {
 
-        @Specialization(guards = "libPattern.isRubyString(pattern)", limit = "1")
-        static RubyRegexp initialize(Object pattern, int options,
+        static final InlinedBranchProfile UNCACHED_BRANCH_PROFILE = InlinedBranchProfile.getUncached();
+
+        @Specialization(
+                guards = {
+                        "libPattern.isRubyString(pattern)",
+                        "patternEqualNode.execute(node, libPattern, pattern, cachedPattern, cachedPatternEnc)",
+                        "options == cachedOptions" },
+                limit = "getDefaultCacheLimit()")
+        static RubyRegexp fastCompiling(Object pattern, int options,
+                @Cached @Shared TruffleString.AsTruffleStringNode asTruffleStringNode,
+                @Cached @Shared RubyStringLibrary libPattern,
+                @Cached("asTruffleStringUncached(pattern)") TruffleString cachedPattern,
+                @Cached("libPattern.getEncoding(pattern)") RubyEncoding cachedPatternEnc,
+                @Cached("options") int cachedOptions,
+                @Cached StringHelperNodes.EqualSameEncodingNode patternEqualNode,
+                @Bind("this") Node node,
+                @Cached("compile(pattern, options, node, libPattern, asTruffleStringNode, UNCACHED_BRANCH_PROFILE)") RubyRegexp regexp) {
+            return regexp;
+        }
+
+        @Specialization(replaces = "fastCompiling", guards = "libPattern.isRubyString(pattern)")
+        RubyRegexp slowCompiling(Object pattern, int options,
                 @Cached InlinedBranchProfile errorProfile,
-                @Cached TruffleString.AsTruffleStringNode asTruffleStringNode,
-                @Cached RubyStringLibrary libPattern,
-                @Bind("this") Node node) {
+                @Cached @Shared TruffleString.AsTruffleStringNode asTruffleStringNode,
+                @Cached @Shared RubyStringLibrary libPattern,
+                @Cached PerformanceWarningNode performanceWarningNode) {
+            performanceWarningNode.warn(
+                    "unbounded creation of regexps causes deoptimization loops which hurt performance significantly, avoid creating regexps dynamically where possible or cache them to fix this");
+            return compile(pattern, options, this, libPattern, asTruffleStringNode, errorProfile);
+        }
+
+        public RubyRegexp compile(Object pattern, int options, Node node, RubyStringLibrary libPattern,
+                TruffleString.AsTruffleStringNode asTruffleStringNode, InlinedBranchProfile errorProfile) {
             var encoding = libPattern.getEncoding(pattern);
             try {
                 return RubyRegexp.create(