+++ src/main/java/org/truffleruby/extra/TruffleGraalNodes.java
@@ -163,7 +163,7 @@ public abstract class TruffleGraalNodes {
     }
 
     @Primitive(name = "assert_not_compiled")
-    public abstract static class AssertNotCompilationConstantNode extends PrimitiveNode {
+    public abstract static class AssertNotCompiledNode extends PrimitiveNode {
 
         @Specialization
         Object assertNotCompiled() {